requirements.txt
requests==2.24.0
fake_headers==1.0.2
numpy==1.17.4
pyproj==2.6.1.post1
mercantile==1.2.1
geopandas==0.8.0
PyVirtualDisplay==2.1
selenium==3.141.0
cloudscraper==1.2.58
selenium_wire==2.1.2
lxml==4.5.1
Shapely==1.7.0
webdriver_manager==2.5.3
python-dotenv==0.19.2
rosreestr2coord==4.0.19
